Added video support and gfycat support.

+++ ImgurBot.py
@@ -1,15 +1,11 @@
 from __future__ import print_function
 from imgurpython import ImgurClient
 from imgurpython.helpers import error
-import os
-from clarifai_system import NSFWDetector
-import VideoAnalyzer
-import urllib
-
+import os, urllib, operator
 
 class Bot:
 
-    def __init__(self):
+    def __init__(self,  videobot, slave_bot):
         """
         Initializes the Imgur Bot with credentials stored in environment variables.
         :return: ImgurClient object
@@ -18,6 +14,8 @@ class Bot:
         IMGUR_CLIENT_SECRET = os.environ.get("IMGUR_CLIENT_SECRET")
         self.client = ImgurClient(IMGUR_CLIENT_ID, IMGUR_CLIENT_SECRET)
         self.supported_video_formats = ['gif','gifv', 'webm', 'mp4']
+        self.slave_bot = slave_bot
+        self.video_bot = videobot
 
     def handle_album(self, album_link):
 
@@ -38,40 +36,50 @@ class Bot:
 
             imgur_flag = album.nsfw
 
-            images_list = self.client.get_album_images(album_id)
-
-            links = [item.link for item in images_list[0:10]
-                     if item.type.split('/')[-1] not in self.supported_video_formats]
-            links_videos = [item.link for item in images_list[0:10]
-                            if item.type.split('/')[-1] in self.supported_video_formats]
-            # Ensures only 10 images/gifs are processed in case album is very large.
-
-            temp1, _ = self.handle_videos(links_videos)
-            temp2, _ = self.handle_images(links)
-
-            status.update(temp1)
-            status.update(temp2)
 
             if imgur_flag:
                 message = 'Album marked NSFW on Imgur.'
                 message = '**[Hover to reveal](#s "' + message + '")**'  # reddit spoiler tag added.
 
             elif not imgur_flag:
-                nsfw_scores = []
-                max_nsfw_score_percentage = 0
 
-                if len(status.values()) > 0:
-                    for k, v in status.items():
-                        nsfw_scores.append(v['nsfw'])
-                    max_nsfw_score_percentage = max(nsfw_scores)
+                images_list = self.client.get_album_images(album_id)
 
-                message = ''
+                links = [item.link for item in images_list[0:10]
+                         if item.type.split('/')[-1] not in self.supported_video_formats]
+                links_videos = [item.link for item in images_list[0:10]
+                                if item.type.split('/')[-1] in self.supported_video_formats]
+                # Ensures only 10 images/gifs are processed in case album is very large.
 
-                if max_nsfw_score_percentage > 50:
-                    message += 'NSFW album. I\'m {0:.2f}% confident.'.format(max_nsfw_score_percentage)
+                temp1, _ = self.handle_videos(links_videos)
+                temp2, _ = self.handle_images(links)
 
-                elif max_nsfw_score_percentage <= 50:
-                    message += 'SFW album. I\'m {0:.2f}% confident.'.format(100 - max_nsfw_score_percentage)
+                status.update(temp1)
+                status.update(temp2)
+
+                # for all images, if SFW - mark SFW.
+                # if any image is not SFW, find out which one.
+
+                max_nsfw= (None, 0)
+                min_sfw = (None, 100)
+                for k,v in status.items():
+                    labels = sorted(status[k].items(), key=operator.itemgetter(1), reverse=True)
+
+                    tag, confidence = labels[0]
+
+                    if tag is 'SFW' and confidence<=min_sfw[1]:
+                        min_sfw = labels[0]
+
+                    elif tag is not 'SFW' and confidence>max_nsfw[1]:
+                        max_nsfw = labels[0]
+
+                if max_nsfw != (None, 0):
+                    message = "Album has "+str(max_nsfw[0])+" image(s). I'm {0:.2f}% confident.".format(max_nsfw[1])
+
+                elif max_nsfw == (None, 0):
+
+                    message = "Album has "+str(min_sfw[0])+" image(s). I'm {0:.2f}% confident.".format(min_sfw[1])
+
 
                 message = '**[Hover to reveal](#s "'+message+' ")**'  #reddit spoiler tag added.
 
@@ -85,21 +93,18 @@ class Bot:
     def handle_images(self, links):
         status = {}
         message = None
-        for each_url in links:
-            link = self.ensure_extension(each_url)
-            if link.split('.')[-1].lower() not in ['gif', 'gifv', 'mp4', 'webm']:
-                status = NSFWDetector().get_predictions([link])
 
-        if len(links) == 1:
-            nsfw_score = [v['nsfw'] for k,v in status.items()][0]
-            sfw_score  = [v['sfw'] for k,v in status.items()][0]
+        valid_links = [self.ensure_extension(aLink) for aLink in links
+                       if aLink.split('.')[-1].lower() not in ['gif', 'gifv', 'mp4', 'webm']]
 
-            if nsfw_score > sfw_score:
-                message = 'NSFW.  I\'m {0:.2f}% confident.'.format(nsfw_score)
-                message = '**[Hover to reveal](#s "' + message + ' ")**'  # reddit spoiler tag added.
-            else:
-                message = 'SFW. I\'m {0:.2f}% confident.'.format(sfw_score)
-                message = '**[Hover to reveal](#s "' + message + ' ")**'  # reddit spoiler tag added.
+        status = self.slave_bot.analyze(valid_links)
+
+        if len(valid_links) == 1:
+            link = valid_links[0]
+            labels = sorted(status[link].items(), key=operator.itemgetter(1), reverse=True)
+
+            message = labels[0][0] + ". I'm  {0:.2f}% confident.".format(labels[0][1])
+            message = '**[Hover to reveal](#s "' + message + ' ")**'  # reddit spoiler tag added.
 
         return status, message
 
@@ -158,19 +163,17 @@ class Bot:
             filename = video_id+'.mp4'
             mp4_link = 'http://i.imgur.com/'+filename
             urllib.urlretrieve(mp4_link, filename)
-            status.update({each_url: VideoAnalyzer.make_prediction(filename)})
-            os.remove(filename)
+            status.update({each_url:self.video_bot.make_prediction(filename)})
+
+            if os.path.exists(filename):
+                os.remove(filename)
 
         if len(links) == 1:
-            nsfw_score = [v['nsfw'] for k,v in status.items()][0]
-            sfw_score  = [v['sfw'] for k,v in status.items()][0]
-
-            if nsfw_score > sfw_score:
-                message = 'NSFW.  I\'m {0:.2f}% confident.'.format(nsfw_score)
-                message = '**[Hover to reveal](#s "' + message + ' ")**'  # reddit spoiler tag added.
-            else:
-                message = 'SFW. I\'m {0:.2f}% confident.'.format(sfw_score)
-                message = '**[Hover to reveal](#s "' + message + ' ")**'  # reddit spoiler tag added.
+            link = links[0]
+            labels = sorted(status[link].items(), key=operator.itemgetter(1), reverse=True)
+            tag, confidence = labels[0]
+            message = tag + ". I'm  {0:.2f}% confident.".format(confidence)
+            message = '**[Hover to reveal](#s "' + message + ' ")**'  # reddit spoiler tag added.
 
         return status, message
 
@@ -178,6 +181,7 @@ class Bot:
         temp = url.split('/')[-1]  # will be <image_id>.<extension> or <image_id>
         if '.' not in temp:
             image_id = temp
+
             url = self.client.get_image(image_id).link
             return url
         else:
@@ -186,4 +190,3 @@ class Bot:
 
 if __name__ == '__main__':
     bot = Bot()
-
